refactor: extract shared utilities from wizard components
Extract duplicated code across the four wizard components and PlaceEditor:
- throttle() → src/lib/throttle.ts (was duplicated 4x)
- isGoogleMapsUrl() → src/lib/google-maps-url.ts (was duplicated 3x)
- MapPinPicker → src/components/admin/MapPinPicker.tsx (map init, draggable
  marker, reverse geocode — was duplicated in PlaceWizard, CommunityWizard,
  and PlaceEditor)
- useWizardSkips() + buildCelebrateUrl() → wizard-helpers.ts (was duplicated
  in all four wizards)

Also removes unused props from PlaceWizard (categories, organizers,
mediaLocations) and the corresponding prop passing in place-new.astro.

Net -347 lines across 7 modified files.
